% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_education_block.R
\name{print_education_block}
\alias{print_education_block}
\title{Print education block}
\usage{
print_education_block(x)
}
\arguments{
\item{x}{data frame with end_date, degree, and institution}
}
\value{
Nothing (invisible NULL)
}
\description{
Print education block
}
