% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_experience_block.R
\name{print_experience_block}
\alias{print_experience_block}
\title{Print experience block}
\usage{
print_experience_block(x)
}
\arguments{
\item{x}{dataframe with columns: enddate, positiontitle, institution, location, startdate, descriptionbullets, and skilltag.}
}
\value{
Nothing (invisible NULL)
}
\description{
Formats and prints a latex string
}
