% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_project_block.R
\name{print_project_block}
\alias{print_project_block}
\title{Print project block}
\usage{
print_project_block(x)
}
\arguments{
\item{x}{data frame with the following columns: project_name, dates, repo, link, description, and tags.}
}
\value{
Nothing (invisible NULL)
}
\description{
Print project block
}
