% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_referee_block.R
\name{print_referee_block}
\alias{print_referee_block}
\title{Print referee block}
\usage{
print_referee_block(x)
}
\arguments{
\item{x}{data frame with columns: name, title, institute, email, and phone}
}
\value{
Nothing (invisible NULL)
}
\description{
Print referee block
}
