% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_skills_block.R
\name{print_skills_block}
\alias{print_skills_block}
\title{Print skills block}
\usage{
print_skills_block(x)
}
\arguments{
\item{x}{dataframe with columns id and text}
}
\value{
None (invisible NULL)
}
\description{
Formats and prints a latex string.
}
