% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto_yaac.R
\name{use_quarto_yaac}
\alias{use_quarto_yaac}
\title{Use the Quarto-YAAC template}
\usage{
use_quarto_yaac(file_name = NULL)
}
\arguments{
\item{file_name}{Name of new qmd file and sub-directory to be created}
}
\value{
a message if extension was successfully copied over
}
\description{
Adds a subdirectory and necessary files to use the Quarto-YAAC
\url{https://github.com/mps9506/quarto-yaac} template.
}
